// resources/[standalone]/illenium-appearance/web/src/components/Appearance/interfaces.ts
export interface PedProp {
  prop_id: number;
  drawable: number;
  texture: number;
}
// ...
interface BlacklistSettings {
  drawables: number[];
  textures: number[];
}

export interface PropSettings {
  prop_id: number;
  drawable: {
    min: number;
    max: number;
  };
  texture: {
    min: number;
    max: number;
  };
  blacklist: BlacklistSettings;
}
// ...
export interface PropConfig {
  hats: boolean;
  glasses: boolean;
  ear: boolean;
  watches: boolean;
  bracelets: boolean;
}